dao metrics in layout
